ocaml bindings to QML properties
