+++ lablqml/_oasis
@@ -10,9 +10,9 @@ Library lablqml
   Path:         .
   BuildTools:   ocamlbuild
   Modules:      QmlContext
-  DataFiles:    stubs.h,kamlo.h
-  CSources:     stubs.c,stubs.h,kamlo.h,modelindex.c,startup_stubs.c,propMap_stubs.c,
-		qml-bindings/property_ml.c
+  DataFiles:    stubs.h,kamlo.h,variant.h
+  CSources:     stubs.c,stubs.h,kamlo.h,modelindex.c,variant.h,startup_stubs.c,propMap_stubs.c,
+		stubs/property_ml.c
   CCOpt:        -std=c++11 -x c++ `pkg-config --cflags Qt5Widgets Qt5Quick` -O2 -Dprivate=public
 
 Executable "test1"
