--- lablqml/qml-bindings/object.h
@@ -1,1 +1,0 @@
-#include <QObject>
