+++ lablqml/stubs/lablqml.h
@@ -1,5 +1,4 @@
-#ifndef STUBS_H
-#define STUBS_H
+#pragma once
 
 extern "C" {
 #include <caml/mlvalues.h>
@@ -23,4 +22,12 @@ extern QModelIndex make_qmodelindex4(int, int, void*, const QAbstractItemModel *
 
 #define Val_none Val_int(0)
 
-#endif // STUBS_H
+static value Val_some(value v) {
+  CAMLparam1(v);
+  CAMLlocal1(some);
+
+  some = caml_alloc(1, 0);
+  Store_field(some, 0, v);
+
+  CAMLreturn(some);
+}
