+++ lablqml/stubs/variant.h
@@ -1,0 +1,6 @@
+#include <lablqml.h>
+#include <QVariant>
+
+extern "C" {
+  value Val_QVariant(value _dest, const QVariant& var);
+}
